Remove startness as it was moved to the specification
