+++ pom.xml
@@ -34,7 +34,7 @@
 
   <properties>
     <version.eclipse.microprofile.config>2.0</version.eclipse.microprofile.config>
-    <version.eclipse.microprofile.health>3.0</version.eclipse.microprofile.health>
+    <version.eclipse.microprofile.health>3.1-SNAPSHOT</version.eclipse.microprofile.health>
     <version.jakarta.servlet>4.0.4</version.jakarta.servlet>
     <version.smallrye-config>2.3.0</version.smallrye-config>
     <version.smallrye-common>1.6.0</version.smallrye-common>
